Move all HostVm-related bits into their own module
Put HostVmLoader and HostVm in their own host_vm module. This creates
cleaner separation between managing of VMs and host-specific bits like
host exit handling.
